% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{fill_mapping}
\alias{fill_mapping}
\title{FUNCTION_TITLE}
\usage{
fill_mapping(df_mapping, dim_cols)
}
\arguments{
\item{df_mapping}{PARAM_DESCRIPTION}

\item{dim_cols}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
data(test_mapping)
list_mapping <- fill_mapping(test_mapping)
print(list_mapping)
}
\seealso{
\code{\link[data.table]{transpose}}
\code{\link[tidyr]{fill}}
\code{\link[tidyselect]{everything}}
}
