% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_all_dimensions_from_mapping_list}
\alias{get_all_dimensions_from_mapping_list}
\title{FUNCTION_TITLE}
\usage{
get_all_dimensions_from_mapping_list(list_mapping, only_subdimensions = FALSE)
}
\arguments{
\item{list_mapping}{PARAM_DESCRIPTION}

\item{only_subdimensions}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[purrr]{vec_depth}}
}
