% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_dimension_mapping_from_df}
\alias{get_dimension_mapping_from_df}
\title{FUNCTION_TITLE}
\usage{
get_dimension_mapping_from_df(
  df_mapping,
  dim_cols = c("scale", "dimension-1", "dimension-2", "dimension-3")
)
}
\arguments{
\item{df_mapping}{PARAM_DESCRIPTION}

\item{dim_cols}{PARAM_DESCRIPTION, Default: c("scale", "dimension-1", "dimension-2", "dimension-3")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
data(test_mapping)
list_mapping <- get_dimension_mapping_from_df(
  test_mapping,
  dim_cols = c('scale', 'dimension-1', 'dimension-2')
)
str(list_mapping)
}
\seealso{
\code{\link[rlist]{list.parse}}, \code{\link[rlist]{list.group}}
\code{\link[purrr]{map}}, \code{\link[purrr]{modify}}
}
