% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreDimension}
\alias{scoreDimension}
\title{FUNCTION_TITLE}
\usage{
scoreDimension(data, list_mapping, how = base::rowMeans)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{list_mapping}{PARAM_DESCRIPTION}

\item{how}{PARAM_DESCRIPTION, Default: mean}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
data(test_data, test_mapping)
df_questionnarie <- test_data; df_mapping <- test_mapping
dim_cols <- c("scale", 'dimension-1', 'dimension-2')
df_questionnarie <- scoreItem(df_questionnarie,
                              df_mapping,
                              dim_cols = dim_cols,
                              item_col = 'id')
list_mapping <- get_dimension_mapping_from_df(df_mapping, dim_cols = dim_cols)
df_questionnarie <- scoreDimension(df_questionnarie, list_mapping)
}
\seealso{
\code{\link[purrr]{vec_depth}}, \code{\link[purrr]{partial}}, \code{\link[purrr]{map_if}}
\code{\link[rlist]{list.ungroup}}
}
