% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreItem}
\alias{scoreItem}
\title{FUNCTION_TITLE}
\usage{
scoreItem(
  data,
  df_mapping,
  dim_cols = c("dimension-1", "dimension-2"),
  item_col = "序号",
  item_type_col,
  key_answer_col,
  NP_col,
  how = base::rowMeans
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{df_mapping}{PARAM_DESCRIPTION}

\item{dim_cols}{PARAM_DESCRIPTION, Default: c("dimension-1", "dimension-2")}

\item{item_col}{PARAM_DESCRIPTION, Default: '序号'}

\item{how}{PARAM_DESCRIPTION, Default: mean}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
data(test_data, test_mapping)
df_questionnarie <- test_data; df_mapping <- test_mapping
df_questionnarie <- scoreItem(
  df_questionnarie,
  df_mapping,
  dim_cols = c("scale", 'dimension-1', 'dimension-2'),
  item_col = 'id',
  NP_col = 'np'
)
}
\seealso{
\code{\link[purrr]{partial}},\code{\link[purrr]{map}}
}
