% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{scoreQuestionnaire}
\alias{scoreQuestionnaire}
\title{FUNCTION_TITLE}
\usage{
scoreQuestionnaire(
  data,
  df_mapping,
  dim_cols = c("scale", "dimension-1", "dimension-2", "dimension-3"),
  item_col = id,
  ...
)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{df_mapping}{PARAM_DESCRIPTION}

\item{dim_cols}{PARAM_DESCRIPTION, Default: c(scale, dimension - 1, dimension - 2, dimension - 3)}

\item{item_col}{PARAM_DESCRIPTION, Default: id}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
scoreQuestionnaire(
  data = test_data,
  df_mapping = test_mapping,
  dim_cols = c("scale", "dimension-1", "dimension-2"),
  item_col = 'id'
)

scoreQuestionnaire(
  data = questionnaire_mixed,
  df_mapping = mapping_mixed,
  dim_cols = c("dimension-1", "dimension-2", "dimension-3"),
  item_col = '序号',
  item_type_col = '题型',
  key_answer_col = '正确答案'
)
}
