% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{fill_mapping}
\alias{fill_mapping}
\title{fill empty mapping dataframe}
\usage{
fill_mapping(df_mapping, dim_cols)
}
\arguments{
\item{df_mapping}{a dataframe indicating the dimension structure}

\item{dim_cols}{column names in \code{df_mapping} indicating the hierarchy of dimensions}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
fill empties in a mapping dataframe with the higher level dimensions
}
\details{
DETAILS
}
\examples{
data(test_mapping)
list_mapping <- fill_mapping(test_mapping)
print(list_mapping)
}
